requirements.txt
pymongo>=4.6.0
redis>=5.0.0
Faker>=24.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
plotly>=5.0.0
pandas>=2.0.0
